pebble: fix compilation failure and atomic fields on 32bits
This commit addresses a few issues on 32bits including some
compilation failure because of int32 overflow and atomic fields
access. According to Go's issue: **On 32 bit platforms, only
64-bit aligned fields can be atomic**. So all atomic uint64/int64
fields should be aligned properly.

+++ batch.go
@@ -798,7 +798,7 @@ func (b *Batch) countData() []byte {
 
 func (b *Batch) grow(n int) {
 	newSize := len(b.data) + n
-	if newSize >= maxBatchSize {
+	if uint64(newSize) >= maxBatchSize {
 		panic(ErrBatchTooLarge)
 	}
 	if newSize > cap(b.data) {
